+++ packages/core-web/CHANGELOG.md
@@ -1,5 +1,18 @@
 # livepeer
 
+## 5.1.0
+
+### Minor Changes
+
+- [#596](https://github.com/livepeer/ui-kit/pull/596) [`3708b73`](https://github.com/livepeer/ui-kit/commit/3708b73eaa72566c531eba07ecc264fc37adb8ac) Thanks [@gioelecerati](https://github.com/gioelecerati)! - Added mirrored broadcast and peer connection error callback
+
+## 5.0.1
+
+### Patch Changes
+
+- Updated dependencies [[`0e5a036`](https://github.com/livepeer/ui-kit/commit/0e5a0365fe9d685271bc96ff5d750ee19fe6406b)]:
+  - @livepeer/core@3.2.8
+
 ## 5.0.0
 
 ### Major Changes